Advance task 1

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	int i, n = 612852475143;
+	unsigned long int i, n = 612852475143;
 
 	for (i = 3; i < 782849; i = 1 + 2)
 	{
